+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## [0.2.0] 2023-10-07
+
+## Added
+
+- Support for Json switch on class attribute to output as JSON
+- Support for using properties in the Title
+
+## Changed
+
+- Quote regex type formats
+- Change value of a null object from `<null>` to `null`
+
 ## [0.1.2-prerelease] 2023-09-05
 
 ## Added